Restructured CLI to be an umbrella application
