+++ apps/cli/lib/electric.ex
