+++ apps/cli/lib/electric/cli/formatting.ex
