+++ apps/cli/lib/electric/commands/accounts.ex
