+++ apps/cli/lib/electric/commands/apps.ex
