+++ apps/cli/lib/electric/commands/auth.ex
