+++ apps/cli/lib/electric/commands/config.ex
