+++ apps/cli/lib/electric/commands/config/init.ex
