+++ apps/cli/lib/electric/config.ex
