+++ apps/cli/lib/electric/password.ex
