+++ apps/cli/lib/electric/progress.ex
