+++ apps/cli/lib/electric/session.ex
