+++ apps/cli/rel/env.bat.eex
