+++ apps/cli/rel/env.sh.eex
