+++ apps/cli/test/electric/cli_test.exs
