+++ apps/cli/test/electric/client_test.exs
@@ -8,7 +8,7 @@ defmodule Electric.ClientTest do
   describe "base_url/0" do
     test "should default to the value given in compilation" do
       assert Client.base_url() ==
-               Application.fetch_env!(:electric_sql_cli, :default_console_url) <> "/api/v1/"
+               Application.fetch_env!(:cli, :default_console_url) <> "/api/v1/"
     end
 
     test "should use the value from the environment if set" do
