+++ apps/cli/test/electric/command_migrations_test.exs
