+++ apps/cli/test/electric/commands/config_test.exs
