+++ apps/cli/test/electric/lexer_test.exs
