+++ apps/cli/test/electric/session_test.exs
