+++ apps/cli/test/support/migration.sql
