+++ apps/cli/test/support/migration2.sql
